Add platform info to --version commands
The output of the --version command to both BDiff and BPatch now
includes information about the platform for which the programs were
compiled.

(Windows XX bit) is appended to the former output, where XX is 32 for
Windows 32 bit builds and 64 for 64 bit builds.

A fatal compiler error is now reported if there is an attempt to compile
the code for non-Windows platforms.